moved shit back to scenesetup
