+++ index.js
@@ -1,7 +1,7 @@
 import Renderer from "./Renderer/Renderer.js";
-import preloadAssets, { bestShader, player, scene } from "./preload.js";
+import preloadAssets, { bestShader } from "./preload.js";
 import Camera from "./Camera.js";
-import sceneSetup from "./sceneSetup.js";
+import { player, scene } from "./sceneSetup.js";
 import { sceneSetTextures, Lights, Gravity } from "./sceneSetup.js";
 import { cameraControlDebug } from "./Camera.js";
 
@@ -31,9 +31,6 @@ function setup() {
   cam.setPosition(0, 0, 0);
   cam.lookAt(0, 0, 1);
   cameraC = new Camera(cam);
-  sceneSetTextures();
-  sceneSetup();
-
   player.setCamera(cameraC);
   noStroke();
 }
