+++ preload.js
@@ -1,25 +1,11 @@
 import TextureBuffer from "./graphics/Texture.js";
 import { loadMap } from "./map.js";
 import GameObject from "./gameObjects/GameObject.js";
-import Player from "./gameObjects/Player.js";
-import Mesh from "./shapes/Mesh.js";
-import Scene from "./Scene.js";
 import Vector3 from "./structs/Vector3.js";
-import MeshCuboid from "./shapes/TestShapes/MeshCuboid.js";
+import { player, scene, setupMeshes, floorMesh } from "./sceneSetup.js";
 
 export const Textures = new TextureBuffer();
 export let bestShader;
-
-let steveMesh;
-export let ghostMesh;
-export let trophyMesh;
-export let player;
-export const scene = new Scene([]);
-export const floorMesh = new MeshCuboid(300, 40, 300);
-
-function loadErrorCallback(err) {
-  console.log("error loading image");
-}
 
 function loadTextures() {
   Textures.addTexture(loadImage("/assets/textures/bricks.png"), "bricks");
@@ -40,16 +26,8 @@ export default async function preloadAssets() {
 
   bestShader = loadShader("./shader/vertex.vert", "./shader/fragment.frag");
 
-  steveMesh = Mesh.createFromObj("./assets/game3dModels/steve.obj");
-  ghostMesh = Mesh.createFromObj("./assets/game3dModels/ghost.obj");
-  trophyMesh = Mesh.createFromObj("./assets/game3dModels/trophy.obj");
+  setupMeshes();
 
-  steveMesh.setTexture(Textures["steve"]);
-  ghostMesh.setTexture(Textures["ghost"]);
-  trophyMesh.setTexture(Textures["trophy"]);
-  floorMesh.setTexture(Textures["bricks"]);
-
-  player = new Player();
   const gObject6 = new GameObject(new Vector3(0, 25, 0), floorMesh);
   gObject6.immovable = true;
   scene.addObjects([gObject6, player]);
